+++ server/server_consume.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"context"
